+++ src/idtrackerai_start_app/arg_parser.py
@@ -186,10 +186,10 @@ def get_parser(defaults: dict | None = None) -> ArgumentParser:
     )
     add_argument(
         "number_of_parallel_workers",
-        "Maximum number of jobs to parallelize segmentation and "
-        "identification image creation. A negative value means using the number "
-        "of CPUs in the system minus the specified value. Zero means using half "
-        "of the number of CPUs in the system. One means no multiprocessing at all",
+        "Maximum number of jobs to parallelize segmentation and identification"
+        " image creation. A negative value means using the number of CPUs in the"
+        " system minus the specified value. Zero means using half of the number of"
+        " CPUs in the system (limited to 8). One means no multiprocessing at all",
         type=int,
     )
     add_argument(